% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_circle.R
\name{add_circle}
\alias{add_circle}
\title{Add Circle}
\usage{
add_circle(data, var, color, group)
}
\arguments{
\item{var}{varaible}

\item{color}{color}

\item{group}{group}

\item{map}{map}
}
\value{

}
\description{
add circle
}
