% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_menu.R
\name{add_menu}
\alias{add_menu}
\title{Add Menu}
\source{
You can see lists of all available icons here:
\url{https://fontawesome.com/icons?from=io}
}
\usage{
add_menu(x, y, z)
}
\arguments{
\item{x}{Sidebar menu which be used for a Shiny input value}

\item{y}{The name of a tab that this menu item will activate}

\item{z}{The available icons}
}
\description{
Add menu for shiny dashboard
}
\examples{
\dontrun{
add_menu("Map", tabName = "Map", icon = icon("map-marked-alt")
}

}
