% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus_world_province.R
\docType{data}
\name{df_daily}
\alias{df_daily}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Daily and Cumulative Cases.}
\format{
A data frame with 9 variables.
\describe{
\item{date}{Date in YYYY-MM-DD format. The date from 2020-01-22 to 2020-10-07. }
\item{confirmed}{Daily new confirmed cases. }
\item{death}{Daily new mortality cases. }
\item{recovered}{Daily new recovered cases. }
\item{active}{Daily new active cases, which could calculated by 'active cases = confirmed - deaths - recoveries'.}
\item{confirmed_cum}{Cumulative confirmed cases from 2020-01-22 to 2020-10-07.}
\item{death_cum}{Cumulative death cases from 2020-01-22 to 2020-10-07.}
\item{recovered_cum}{Cumulative recovered cases from 2020-01-22 to 2020-10-07.}
\item{active_cum}{Cumulative active cases from 2020-01-22 to 2020-10-07, which could calculated by 'cumulative active cases = cumulative confirmed - cumulative deaths - cumulative recoveries'}
}
}
\source{
Johns Hopkins University Center for Systems Science and Engineering
(JHU CCSE) Coronavirus \href{https://systems.jhu.edu/research/public-health/ncov/}{website}.
}
\usage{
df_daily
}
\description{
Daily new confirmed, recovered, death and active cases with their cumulative cases from 2020-01-22 to 2020-10-07.
}
\details{
The dataset contains the daily summary of Coronavirus cases
(confirmed, death, and recovered), in global.
}
\keyword{COVID19}
\keyword{coronavirus}
\keyword{datasets}
\keyword{in}
\keyword{the}
\keyword{world}
