% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus_world_province.R
\docType{data}
\name{df_tree}
\alias{df_tree}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) Dataset for each country}
\format{
A data frame with 3 variables.
\describe{
\item{country}{Countries}
\item{type}{Includes confirmed, recovered, death and active cases}
\item{total}{Cumulative amount until 2020-10-07}
}
}
\source{
Johns Hopkins University Center for Systems Science and Engineering
(JHU CCSE) Coronavirus \href{https://systems.jhu.edu/research/public-health/ncov/}{website}.
}
\usage{
df_tree
}
\description{
Cumulative Confirmed, Death, Recovered and Active Cases Dataset in Countries and Province/States
}
\keyword{datasets}
