% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_table.R
\name{filter_table}
\alias{filter_table}
\title{Filter df table}
\usage{
filter_table(df, country_name, type_name, date1, date2)
}
\arguments{
\item{df}{Enter your origin data set.}

\item{country_name}{Countries you want to selected.}

\item{type_name}{Select different type cases in the dataframe, which includes confirmed, recovered and deaths.}

\item{date1}{The starting point of the time period.}

\item{date2}{The ending point of the time period.}
}
\description{
This function aim to filter different countries, type of cases, time period for the raw data.
}
\examples{
df_Canada <- filter_table(df, "Canada",
"confirmed", lubridate::ymd("2020-01-22"), lubridate::ymd("2020-01-31"))
tibble::view(df_Canada)

}
