% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_dataframe.R
\name{group_dataframe}
\alias{group_dataframe}
\title{Summarize Data Frame}
\usage{
group_dataframe(df, groupBy)
}
\arguments{
\item{df}{the data called 'df'}

\item{groupBy}{group by each columns}
}
\description{
summarize data and transfer it into dataframe.
}
