% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus_world_province.R
\docType{data}
\name{recent}
\alias{recent}
\title{The 2019 Novel Coronavirus COVID-19 (2019-nCoV) in 2020-10-07.}
\format{
A data frame with 8 variables.
\describe{
\item{date}{The Most Recent Date in YYYY-MM-DD format in 2020-10-07.}
\item{country}{Name of country/region.}
\item{lat}{Latitude of center of geographic region, defined as either
country or, if available, province.}
\item{long}{Longitude of center of geographic region, defined as either
country or, if available, province.}
\item{province}{Name of province/state, for countries where data is
provided split across multiple provinces/states.}
\item{confirmed}{Confirmed Cases in 2020-10-07.}
\item{death}{Mortality Cases in 2020-10-07.}
\item{recovered}{Recovered Cases in 2020-10-07.}
}
}
\source{
Johns Hopkins University Center for Systems Science and Engineering
(JHU CCSE) Coronavirus \href{https://systems.jhu.edu/research/public-health/ncov/}{website}.
}
\usage{
recent
}
\description{
Daily summary of the Coronavirus (COVID-19) cases by state/province in 2020-10-07.
}
\details{
The dataset contains the most recent summary Coronavirus cases
(confirmed, death, and recovered), by state/province when making this shiny app.
}
\keyword{2020-10-07}
\keyword{COVID19}
\keyword{and}
\keyword{coronavirus}
\keyword{coutries}
\keyword{in}
\keyword{province/states}
