% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider_date.R
\name{slider_date}
\alias{slider_date}
\title{Slider Date}
\usage{
slider_date(inputid, inputDisplay, date)
}
\arguments{
\item{inputid}{The input slot that will be used to access the value}

\item{inputDisplay}{Display label for the control}

\item{date}{The minimum and maximum date can be selected}
}
\value{

}
\description{
This function constructs a slider widget to select the minimum date and maximum date from a range.
}
\examples{
\dontrun{
slider_date("dateInput", "Date:", df$date)
}

}
