% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dataframe.R
\name{summariseData}
\alias{summariseData}
\title{Summarize Data Frame}
\usage{
summariseData(df, groupBy)
}
\arguments{
\item{df}{the data called 'df'}

\item{groupBy}{group by each columns}
}
\description{
summarize data and transfer it into dataframe.
}
