% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_menu.R
\name{add_menu}
\alias{add_menu}
\title{Dashboard Menu}
\source{
You can see lists of all available icons here:
\url{https://fontawesome.com/icons?from=io}
}
\usage{
add_menu(id, text_name, icon_name)
}
\arguments{
\item{id}{Sidebar menu which be used for a Shiny input value}

\item{text_name}{The name of a tab that this menu item will activate}

\item{icon_name}{The available icons}
}
\description{
Add menu for shiny dashboard
}
\examples{
add_menu("Map", "Map", "map-marked-alt")

}
